requirements.txt
pandas>=2.0
openpyxl>=3.1
python-telegram-bot>=21.0
